Modified "entity.py" to have a loading function, and moved "create_character" to "entity.py"
