+++ entity.py
@@ -9,7 +9,7 @@ class Entity:
             char_name (str): Name of character
             char_health (int): Character health
             char_base_dmg (int): Character base damage
-            equipped_item (str): Currently equipped item (has to be an item)
+            equipped_item (str): Currently equipped item (has to be an item class)
         """
         self.inv = inventory
         self.name = char_name
@@ -63,8 +63,9 @@ class Entity:
         """
         self.inv.save_to_file()
         with open("char.txt", "w") as file:
-            file.write(self.name + "," + str(self.health) + "," + str(self.base_dmg) + "," + str(self.equipped_item) + "\n")
-            
+            file.write(
+                self.name + "," + str(self.health) + "," + str(self.base_dmg) + "," + str(self.equipped_item) + "\n")
+
     def use_item(self, item_name: str):
         """Uses an item to heal
 
@@ -89,7 +90,7 @@ class Entity:
                     return -1
             else:
                 return -1
-            
+
     def check_health(self):
         """Checks the health of the character, to see if they are dead
 
@@ -100,7 +101,7 @@ class Entity:
             return -1
         else:
             return 1
-            
+
     def attack(self, enemy: object):
         """Attacks the enemy
 
@@ -115,3 +116,63 @@ class Entity:
             return enemy.check_health()
         else:
             return enemy.check_health()
+
+
+def create_character(inv, char_name, char_health, char_base_dmg):
+    """Creates the character, and has integer checking
+
+    Args:
+        inv (Inventory): Inventory class
+        char_name (str): Name of character
+        char_health (int): Character health
+        char_base_dmg (int): Character base damage
+
+    Returns:
+        Entity: Entity class if successful,
+        int: -1 if failed
+    """
+    if char_health.isdigit() and char_base_dmg.isdigit():
+        char_health = int(char_health)
+        char_base_dmg = int(char_base_dmg)
+        plr = Entity(inv, char_name, char_health, char_base_dmg, "None")
+        plr.save()
+        return plr
+    else:
+        return -1
+
+
+def load_char(inv: Inventory):
+    """Loads the character
+
+    Args:
+        inv (Inventory): Inventory class
+
+    Returns:
+        Entity: Entity class
+    """
+    try:
+        with open("char.txt", "r") as file:
+            char = [line.strip().split(',') for line in file]
+            for char_data in char:
+                name = char_data[0]
+                health = int(char_data[1])
+                base_dmg = int(char_data[2])
+                equipped_item = char_data[3]
+                plr = Entity(inv, name, health, base_dmg, equipped_item)
+                return plr
+            inv.load_from_file()
+    except OSError:
+        temp_run = True
+
+        while temp_run:
+            char_name = input("Enter the name of your character: ")
+            char_health = input("Enter the health of your character: ")
+            char_base_dmg = input("Enter the base damage of your character: ")
+            plr = create_character(inv, char_name, char_health, char_base_dmg)
+            if plr == -1:
+                print("The health and damage must be integers")
+            else:
+                print("Character {} created".format(char_name))
+                plr.save()
+                return plr
+                temp_run = False
